split into separate files pushed most printing methods to printer
